Tighten margins on the login page header; fix exhaustive deps warnings (#89)
* Tighten up content margins

* Fix some React hooks exhaustive dependency warnings

Fixes two exhaustive dependency warnings - one where the currentLocationUuid
string and the hideWelcome message boolean needed to be included in the
dependency array; and another where the onChangeLocation function needed
to be added to the array. One remains from the useEffect that toggles a location
search from the location picker.